+++ src/utils/Utils.h
@@ -9,14 +9,14 @@
  * @param count  количество байт
  * @return int
  */
-int getIntFromBytes(byte* buffer, int count);
+size_t getIntFromBytes(byte* buffer, int count);
 
 /**
  * Функция нужна для получения int из VarInt little-endian числа
  * @param buffer буффер с little-endian числом
- * @return 1. int - число
+ * @return 1. uint16_t - число (8 байт)
  *         2. int - кол-во байт, которое занимает это число
  */
-tuple<int, int> getVarInt(byte* buffer);
+tuple<size_t, int> getVarInt(byte* buffer);
 
 #endif //GENISYS_BLOCK_FROM_UTXO_UTILS_H
